+++ src/modules/Missions/UseCases/DeleteWarningMission/DeleteWarningMissionUseCase.ts
@@ -1,4 +1,5 @@
 import { AppError } from "../../../../shared/errors/AppError"
+import { IInputDeleteWarning, IOutputGenericWarningsMissionDTOS } from "../../dtos/IWarningsMissionDTOS"
 import { WarningsMission } from "../../infra/typeorm/entities/WarningMission"
 import { MenagerPermissionRespository } from "../../infra/typeorm/repositories/MenagerPermissionRepository"
 import { IWarningsMissionRepository } from "../../repositories/IWarningsMissionRepository"
@@ -10,9 +11,11 @@ class DeleteWarningMissionUseCase{
     this.warningsMissionRepository = warningsMissionRepository
     this.menagePermissionMission = new MenagerPermissionRespository()
   }
-  async execute({id,id_agent_token}):Promise<WarningsMission>{
+  async execute({id,id_agent_token}:IInputDeleteWarning):Promise<IOutputGenericWarningsMissionDTOS>{
     if(!id) throw new AppError("Value of warning is undefined.")
-    const allowDeleteWarning = await this.menagePermissionMission.confirmePermissionMission({id_agent_token})
+     const findWarning = await this.warningsMissionRepository.findById(id)
+     if(!findWarning) throw new AppError("Warning not found")
+    const allowDeleteWarning = await this.menagePermissionMission.confirmePermissionMission({id_agent_token,id_mission:findWarning.id_mission})
     if(!allowDeleteWarning) throw new AppError("Agent authentcated haven't autorization to execute that action.")
     const deletedWarnigMission = await this.warningsMissionRepository.delete(id)
     return deletedWarnigMission
